Fix bug when saving blank pages
The problem was that `_page['content']` can be an empty string, and
`lxml.html.fromstring('')` raises `lxml.etree.XMLSyntaxError`, which wasn't
caught. The `lxml.etree.ParserError` is raised when, for example,
`_page['content'] == '    '`.

Basically, if any errors are found while parsing the page to inject views, we
should abort and not inject the views, and not break CKAN. I could've caught
`Exception` instead of specific errors, but I prefered to catch specific errors
to be more explicit.