Refactor theme directory to contain only one templates directory
Instead of having organizations and groups and blog posts and
pages and bases and other bases and more bases in a directory
called templates_main/ckanext_pages and then single extends file
in templates_organization and templates_group directories with
the rest of the related files in the main directory, this moves
files similar in concept into the same directories:

* default - main directory that is always loaded
* default/blog - blog article templates
* default/pages - pages templates
* default/editor - editor templates
* organization_pages - all templates for the configurable org pages
* group_pages - all templates for the configurable group pages

Files are also when possible given names that reflect the
structure of more common ckan templates (index.html instead of
pages_list.html, snippets/helper.html instead of
base_edit_form.html etc.)

With this the overall templates structure should be both more
understandable and more tweakable for each functionality because
blog articles and pages might need to be presented differently
(code should actually allow for this in a better way but that's
for the future).

Oh yes, and this adds breadcrumbs as well (which was the original
intent of this commit but you know, yak-shaving).