Remove extraneous tags added by lxml that prevent proper rendering
If lxml is enabled, the `_inject_views_into_page` method of the
controller does some tweaking of the HTML tree. If the content passed is
a text snippet (like 99.99% of the markdown content passed) it will wrap
it on <p> tags (or if the snippet starts with a <p> tag, with a <div>
tag). This will cause the snippet to not be properly rendered.